refactor: move required version tests from utils into domain test file
